+++ R/sp.plot.R
@@ -1,17 +1,19 @@
 #' @name sp.plot
 #' @title plotting (re)SPECIATE profiles
-#' @aliases sp_plot_profile
+#' @aliases sp_plot_profile sp_plot_species
 
 #' @description General plots for \code{respeciate} objects.
 
 #' @description \code{sp_plot} functions generate plots for supplied
-#' (re)SPECIATE profile data sets.
+#' (re)SPECIATE data sets.
 #' @param x A \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
 #' profiles.
-#' @param id numeric, indices of profiles to use when
-#' plotting (e.g. \code{id=1:6} plots first 6 profiles).
-#' @param multi.profile character, how plot should handle
-#' multiple profiles, e.g. 'group' or 'panel' (default
+#' @param id numeric, the indices of profiles or species to use when
+#' plotting with \code{sp_plot_profile} or \code{sp_plot_species},
+#' respectively. For example, \code{sp_plot_profile(x, id=1:6)} plots
+#' first 6 profiles in \code{respeciate} object \code{x}.
+#' @param multi.profile character, how \code{sp_plot_profile} should
+#' handle multiple profiles, e.g. 'group' or 'panel' (default
 #' group).
 #' @param order logical, order the species in the
 #' profile(s) by relative abundance before plotting.
@@ -20,9 +22,25 @@
 #' plotting functions.
 #' @param silent logical, hide warnings when generating plots (default
 #' \code{FALSE})
-#' @param multi.species, like \code{multi.profile} but for species.
+#' @param multi.species, character, like \code{multi.profile} in
+#' \code{sp_plot_profile} but for species in \code{sp_plot_species}.
 #' @return \code{sp_plot} graph, plot, etc usually as a trellis object.
 #' @note These functions are currently in development, so may change.
+#' @references Most \code{respeciate} plots make extensive use of
+#' \code{lattice} and \code{latticeExtra} code:
+#'
+#' Sarkar D (2008). \emph{Lattice: Multivariate Data Visualization with R}.
+#' Springer, New York. ISBN 978-0-387-75968-5, \url{http://lmdvr.r-forge.r-project.org}.
+#'
+#' Sarkar D, Andrews F (2022). \emph{latticeExtra: Extra Graphical Utilities Based
+#' on Lattice}. R package version 0.6-30,
+#' \url{https://CRAN.R-project.org/package=latticeExtra}.
+#'
+#' They also incorporate ideas from \code{loa}:
+#'
+#' Ropkins K (2023). \emph{loa: various plots, options and add-ins for use with lattice}.
+#' R package version 0.2.48.3, \url{https://CRAN.R-project.org/package=loa}.
+
 
 #functions
 # sp_plot_profile
@@ -30,22 +48,50 @@
 
 # plot.respeciate is wrapper for sp_plot_profile
 
-#use unexported
-# rsp_plot_fix
+#uses unexported code
+#  rsp_plot_fix
+#  rsp_yscale.component.log10 (currently in sp.pls.r)
+
 
 
 
 #JOBS
 #######################
 
-#reference lattice and latticeEXtra packages in documents...
+#references may need formatting tidying
+#   currently these are lattice, latticeEXtra and loa...
+#   check roxygen2 guidance ???
 
 #all functions need work
 #see function fix, tidy, etc job notes in code
+#     ALL need better colour handling for large numbers of cases
+#         typically group handling...
+#            maybe a variation on col=rainbow ???
+
+#examples
+# maybe
+# sp_plot_profile(spq_pm.ae8())
+#   (allows most lattice style plot control, etc key=list(...))
+#   (but includes some short cuts to common handling, e.g. log=T to
+#      log y scales and reformat y axes)
+# sp_plot_profile(spq_pm.ae8(), key=list(space="top", columns=2), log=T)
+
+#color defaults...
+#issue current default wraps if you exceed number of cols in default set.
+#from: https://stackoverflow.com/questions/26314701/r-reducing-colour-saturation-of-a-colour-palette
+#function(x) colorRampPalette(rainbow(12, s = 0.5, v = 1)[2:11],interpolate = "spline")(x)
+##   ?? could extrapolate the default colors using something like above ???
+
+
+
+# dennis asked for data as part of return
+#    that is do-able but may need an object class
+#         (maybe like the openair code...)
 
 #thinking about an sp_plot_compare(x, y)
 #  to compare profile x and profile(s) y
 #  started project (in own-notes)
+
 
 ###################################
 #sp_plot_profile
@@ -71,6 +117,14 @@
 
 #see in code notes about jobs
 
+##############################
+#testing
+#   reset.x as option to change
+#        x access handing
+#        wondering about a general fix
+#             upfront so applied to x (rsp data.frame)
+#               what is x, how is it formatted, etc
+#               then same for y, groups and cond...
 
 sp_plot_profile <-   function(x, id, multi.profile = "group",
                               order=TRUE, log=FALSE, ...,
@@ -286,12 +340,12 @@ sp_plot_profile <-   function(x, id, multi.profile = "group",
 
 #in development
 
-#taken straight from sp_plot_profile
+#lot taken straight from sp_plot_profile
 #so lots of redundancy
 
 sp_plot_species <- function(x, id, multi.species = "group",
-                              order=FALSE, log=FALSE, ...,
-                              silent=FALSE){
+                            order = FALSE, log = FALSE,
+                            ..., silent = FALSE){
 
   #setup
   .x.args <- list(...)
@@ -312,6 +366,10 @@ sp_plot_species <- function(x, id, multi.species = "group",
 
 
   #need to get species as character
+  ##############################
+  #if already factor ???
+  #   user could be forcing order
+  ##############################
   .sp.ord <- as.character(unique(x$SPECIES_ID))
   #.sp.pro <- unique(x$PROFILE_CODE)
   #n/profile handling
@@ -360,20 +418,34 @@ sp_plot_species <- function(x, id, multi.species = "group",
   }
 
   ####################################
-  #like sp_plot_profile...
+  #current species ordering by id arg...
+  #(see below about reordering)
   ####################################
-  species <- unique(x$SPECIES_NAME)
-  #should think about other naming options???
+  species <- species[species %in% unique(x$SPECIES_ID)]
+  x$SPECIES_ID <- factor(x$SPECIES_ID, levels=species)
+  x <- x[order(x$SPECIES_ID),]
+  x$SPECIES_NAME <- factor(x$SPECIES_NAME, unique(x$SPECIES_NAME))
+  species<- levels(x$SPECIES_NAME)
+  sp.ord <- as.numeric(factor(species, levels=sort(species)))
 
-#ignoreing order at moment
+  ##################################
+  #should think about other naming options???
+  ##################################
+  #print(species)
+  #print(sp.ord)
+
+#ignoring option to re-order at moment
 
   #order largest to smallest
   #############################
+  #like to enable this
   #like to also be able to order by molecular weight
+  #need to decide handling if species is already a factor... ???
+  #need to decide if this should work off species_id or species_name... ???
   ##############################
   if(order){
     ################################
-    #bit of a cheat...
+    #taken from _profile plots
     ################################
     test <- x
     test$PROFILE_CODE <- ".default"
@@ -391,16 +463,14 @@ sp_plot_species <- function(x, id, multi.species = "group",
   }
   x <- x[c(".value","PROFILE_CODE", "PROFILE_NAME", "SPECIES_NAME")]
 
-
   ##################
-  #profile bar chart
+  #species trend line plot
   ##################
 
   #dcast and melt to add in any missed entries as NAs
-  #(for the plot trail)
-  #not padding
+  #(to force trend line gaps)
+  #not padding, obviously not dropping nas...
   x <- sp_melt_wide(sp_dcast_species(x), pad=FALSE, drop.nas = FALSE)
-
 
   ###############################
   #species handling
@@ -423,16 +493,35 @@ sp_plot_species <- function(x, id, multi.species = "group",
   #  convert to factor
   #       but then by default lattice shows all factors labels...
   #  format using a supplied function???
-  x$x.id <- as.numeric(factor(x$PROFILE_CODE))
+
+  ############################
+  #could move this top and apply
+  #before plotting??
+
+  if("reset.x" %in% names(.x.args)){
+    #initial test reset.x
+    #  this is a function and it is applied to profile_code
+    #      to build the x axis...
+    x$.x <- .x.args$reset.x(x$PROFILE_CODE)
+    .xlab <- ""
+  } else {
+    x$.x <- as.numeric(factor(x$PROFILE_CODE))
+    .xlab <- "Sample [index]"
+  }
 
 
-  p1.ls <- list(x= .value~x.id,
-                data=x, ylab="Measurement", xlab="Sample [index]",
+  ##############################
+  #species alignment
+
+  p1.ls <- list(x= .value~.x,
+                data=x, ylab="Measurement", xlab=.xlab,
                 type="l",
                 #NB: prepanel seemed to break ylim when stacking
                 panel = function(x, y, ...){
-                  rsp_panelPal("grid", list(h=-1,v=-1, col="grey", lty=3),
-                               panel.grid, ...)
+                  at.x <- pretty(x)
+                  at.y <- pretty(y)
+                  rsp_panelPal("grid", list(h=at.y,v=at.x, col="grey", lty=3),
+                               panel.abline, ...)
                   panel.xyplot(x=x, y=y, ...)
                 },
                 between=list(y=.2),
@@ -449,10 +538,10 @@ sp_plot_species <- function(x, id, multi.species = "group",
   if(length(species)>1){
     if(tolower(multi.species) %in% c("panel", "panels")){
       #paneling multiple panels
-      p1.ls$x <- .value~x.id | SPECIES_NAME
+      p1.ls$x <- .value~.x | SPECIES_NAME
     } else {
       #grouping multiple panels
-      p1.ls$x <- .value~x.id
+      p1.ls$x <- .value~.x
       p1.ls$groups <- x$SPECIES_NAME
     }
   }
@@ -480,8 +569,15 @@ sp_plot_species <- function(x, id, multi.species = "group",
     }
   } else {
     p1.ls$col <- if("groups" %in% names(p1.ls)){
-      rep(trellis.par.get("superpose.line")$col,
-          length.out=length(species))
+      colorRampPalette(rainbow(12, s = 0.5, v = 1),
+                               interpolate = "spline")(length(species))
+      #or:
+      #colorRampPalette(rainbow(12, s = 0.5, v = 1),interpolate = "spline")(x)
+      #was:
+      #colorRampPalette(trellis.par.get("superpose.line")$col,
+      #    interpolate = "spline")(length(species))
+      #rep(trellis.par.get("superpose.line")$col,
+      #    length.out=length(species))
     } else {
       trellis.par.get("superpose.line")$col[1]
     }
@@ -501,7 +597,7 @@ sp_plot_species <- function(x, id, multi.species = "group",
                  #title="Legends",
                  lines=list(col=rep(p1.ls$col,
                                     length.out=length(species))),
-                 text = list(species, cex=0.7))
+                 text = list(levels(x$SPECIES_NAME), cex=0.7))
     p1.ls$key <- if("key" %in% names(p1.ls)){
       modifyList(.tmp, p1.ls$key)
     } else {